Move test_contains.py
